infrastructure/docker/mlflow/requirements.txt
# MLflow Tracking Server Requirements
# Production-ready dependencies with version pinning

# Core MLflow
mlflow==2.8.1
mlflow-skinny==2.8.1

# Database connectors
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Object storage
boto3==1.34.0
botocore==1.34.0
s3fs==2023.12.0

# Cloud providers
azure-storage-blob==12.17.0
google-cloud-storage==2.10.0

# Web framework
flask==2.3.3
gunicorn==21.2.0
waitress==2.1.2

# Data science libraries
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.7.4
seaborn==0.12.2

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.7
authlib==1.2.1

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Performance
uvloop==0.19.0
httptools==0.6.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.0.7 